Update ping.js
Fixed issue where server sends a ping right after the NICK command is sent, causing a timeout on connection, since the regex doesn't match it for some strange reason. Tested to make sure the ping/pong worked by letting the script run for quite a while.

The issue is when that ping comes in, it is in hex:
50 49 4e 47 20 3a 38 43 35 32 35 30 34 34 d

Notice the d by it self. Some how it's truncating the last byte which should be 0d - which would be a CR. This might be an issue with UnrealIRCD (3.x and 4.x).